requirements.txt
streamlit>=1.32.0
pillow>=10.2.0
numpy>=1.26.4
